:truck: move definitions
